rename to EditableText
