--- src/components/atoms/EditableNameText/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./EditableNameText";
