+++ src/components/atoms/EditableText/EditableText.test.tsx
@@ -2,18 +2,18 @@ import React from "react";
 
 import { fireEvent, render, screen } from "@testing-library/react";
 
-import EditableNameText from "./EditableNameText";
+import EditableText from "./EditableText";
 
 describe('default EditableNameText', () => {
   it('displays children', () => {
-    render(<EditableNameText onUpdate={jest.fn()}>Test 123</EditableNameText>);
+    render(<EditableText onUpdate={jest.fn()}>Test 123</EditableText>);
     const component = screen.queryByText("Test 123");
     expect(component).not.toBeNull();
     expect(component).toBeInstanceOf(HTMLSpanElement);
   });
   
   it('changes to text input after doubleclick', () => {
-    render(<EditableNameText onUpdate={jest.fn()}>Test 456</EditableNameText>);
+    render(<EditableText onUpdate={jest.fn()}>Test 456</EditableText>);
     let component = screen.getByText("Test 456");
     expect(component).toBeInstanceOf(HTMLSpanElement);
